Split details file into folder
