+++ pyqt-apps/siriushla/as_rf_control/details/__init__.py
@@ -1,0 +1,7 @@
+from .cavity_status import CavityStatusDetails
+from .fdl import FDLDetails
+from .interlock import LLRFInterlockDetails
+from .sl_err import SlowLoopErrorDetails
+from .sl_param import SlowLoopParametersDetails
+from .temp_monitor import TempMonitor
+from .transm_line import TransmLineStatusDetails
