+++ pyqt-apps/siriushla/as_rf_control/details/cavity_status.py
@@ -1,0 +1,155 @@
+"""Cavity Status Datails."""
+
+from qtpy.QtCore import Qt
+from qtpy.QtWidgets import QFormLayout, QGridLayout, QHBoxLayout, QLabel
+
+from ...widgets import PyDMLedMultiChannel, SiriusDialog, SiriusLabel, \
+    SiriusLedAlert
+from ..util import SEC_2_CHANNELS
+
+
+class CavityStatusDetails(SiriusDialog):
+    """Cavity Status Datails."""
+
+    def __init__(self, parent=None, prefix='', section=''):
+        super().__init__(parent)
+        self.prefix = prefix
+        self.prefix += ('-' if prefix and not prefix.endswith('-') else '')
+        self.section = section.upper()
+        self.chs = SEC_2_CHANNELS[self.section]
+        self.setObjectName(self.section + 'App')
+        self.setWindowTitle(self.section + ' Cavity Detailed Status')
+        self._setupUi()
+
+    def _setupUi(self):
+        lay_temp1 = QFormLayout()
+        lay_temp1.setHorizontalSpacing(9)
+        lay_temp1.setVerticalSpacing(9)
+        lay_temp1.setLabelAlignment(Qt.AlignRight)
+        lay_temp1.setFormAlignment(Qt.AlignTop)
+        lb_temp1 = QLabel('Cell and Coupler\nTemperatures\nPT100', self)
+        lb_temp1.setStyleSheet(
+            'font-weight:bold; qproperty-alignment:AlignCenter;')
+        lay_temp1.addRow(lb_temp1)
+        lims = self.chs['Cav Sts']['Temp']['Cells Limits']
+        tooltip = 'Interlock limits: \nMin: ' + str(lims[0]) + \
+            '°C, Max: '+str(lims[1])+'°C'
+        for idx, cell in enumerate(self.chs['Cav Sts']['Temp']['Cells']):
+            lbl = SiriusLabel(self, self.prefix+cell[0])
+            lbl.showUnits = True
+            lbl.setStyleSheet('min-width:3.5em; max-width:3.5em;')
+            led = PyDMLedMultiChannel(
+                self,
+                {self.prefix+cell[0]: {'comp': 'wt', 'value': lims},
+                 self.prefix+cell[0].replace('T-Mon', 'TUp-Mon'): 0,
+                 self.prefix+cell[0].replace('T-Mon', 'TDown-Mon'): 0})
+            led.setToolTip(tooltip)
+            hbox = QHBoxLayout()
+            hbox.setAlignment(Qt.AlignLeft)
+            hbox.addWidget(lbl)
+            hbox.addWidget(led)
+            lay_temp1.addRow('Cell '+str(idx + 1)+': ', hbox)
+        ch_coup = self.chs['Cav Sts']['Temp']['Coupler'][0]
+        lims_coup = self.chs['Cav Sts']['Temp']['Coupler Limits']
+        lb_coup = SiriusLabel(self, self.prefix+ch_coup)
+        lb_coup.setStyleSheet('min-width:3.5em; max-width:3.5em;')
+        lb_coup.showUnits = True
+        led_coup = PyDMLedMultiChannel(
+            self,
+            {self.prefix+ch_coup: {'comp': 'wt', 'value': lims_coup},
+             self.prefix+ch_coup.replace('T-Mon', 'TUp-Mon'): 0,
+             self.prefix+ch_coup.replace('T-Mon', 'TDown-Mon'): 0})
+        led_coup.setToolTip(
+            'Interlock limits: \n'
+            'Min: '+str(lims_coup[0])+'°C, Max: '+str(lims_coup[1])+'°C')
+        hb_coup = QHBoxLayout()
+        hb_coup.setAlignment(Qt.AlignLeft)
+        hb_coup.addWidget(lb_coup)
+        hb_coup.addWidget(led_coup)
+        lay_temp1.addRow('Coupler: ', hb_coup)
+
+        lay_temp2 = QFormLayout()
+        lay_temp2.setHorizontalSpacing(9)
+        lay_temp2.setVerticalSpacing(9)
+        lay_temp2.setLabelAlignment(Qt.AlignRight)
+        lay_temp2.setFormAlignment(Qt.AlignTop | Qt.AlignHCenter)
+        lb_temp2 = QLabel('Cell\nTemperatures\nThermostats', self)
+        lb_temp2.setStyleSheet(
+            'font-weight:bold; qproperty-alignment:AlignCenter;')
+        lay_temp2.addRow(lb_temp2)
+        for idx, cell in enumerate(self.chs['Cav Sts']['Temp']['Cells']):
+            led = SiriusLedAlert(self)
+            led.setToolTip('Interlock limits:\nMax: 60°C')
+            led.channel = self.prefix+cell[0].replace('T-Mon', 'Tms-Mon')
+            lay_temp2.addRow('Cell '+str(idx + 1)+': ', led)
+
+        lay_dtemp = QFormLayout()
+        lay_dtemp.setHorizontalSpacing(9)
+        lay_dtemp.setVerticalSpacing(9)
+        lay_dtemp.setLabelAlignment(Qt.AlignRight)
+        lay_dtemp.setFormAlignment(Qt.AlignTop | Qt.AlignHCenter)
+        lb_dtemp = QLabel('Disc\nTemperatures\nThermostats', self)
+        lb_dtemp.setStyleSheet(
+            'font-weight:bold; qproperty-alignment:AlignCenter;')
+        lay_dtemp.addRow(lb_dtemp)
+        for idx, disc in enumerate(self.chs['Cav Sts']['Temp']['Discs']):
+            led = SiriusLedAlert(self)
+            led.setToolTip('Interlock limits:\nMax: 60°C')
+            led.channel = self.prefix+disc
+            lay_dtemp.addRow('Disc '+str(idx)+': ', led)
+
+        lay_flwrt = QFormLayout()
+        lay_flwrt.setHorizontalSpacing(9)
+        lay_flwrt.setVerticalSpacing(9)
+        lay_flwrt.setLabelAlignment(Qt.AlignRight)
+        lay_flwrt.setFormAlignment(Qt.AlignTop)
+        lb_flwrf = QLabel('Flow Switches', self)
+        lb_flwrf.setStyleSheet(
+            'font-weight:bold; qproperty-alignment:AlignCenter;')
+        lay_flwrt.addRow(lb_flwrf)
+        for flwsw, pvn in self.chs['Cav Sts']['FlwRt'].items():
+            led = SiriusLedAlert(self, self.prefix+pvn)
+            lay_flwrt.addRow(flwsw+': ', led)
+
+        self.led_couppressure = SiriusLedAlert(
+            self, self.prefix+self.chs['Cav Sts']['Vac']['Coupler ok'])
+        self.led_pressure = SiriusLedAlert(self)
+        self.led_pressure.setToolTip('Interlock limits:\nMax: 5e-7mBar')
+        self.led_pressure.channel = \
+            self.prefix+self.chs['Cav Sts']['Vac']['Cells ok']
+        lay_vac = QFormLayout()
+        lay_vac.setHorizontalSpacing(9)
+        lay_vac.setVerticalSpacing(9)
+        lay_vac.setLabelAlignment(Qt.AlignRight)
+        lay_vac.setFormAlignment(Qt.AlignTop)
+        lb_vac = QLabel('Vacuum', self)
+        lb_vac.setStyleSheet(
+            'font-weight:bold; qproperty-alignment:AlignCenter;')
+        lay_flwrt.addRow(lb_vac)
+        lay_vac.addRow('Pressure Sensor: ', self.led_couppressure)
+        lay_vac.addRow('Vacuum: ', self.led_pressure)
+
+        lbl = QLabel('Cavity - Detailed Status', self)
+        lbl.setStyleSheet(
+            'font-weight:bold; qproperty-alignment:AlignCenter;')
+        lay = QGridLayout(self)
+        lay.setHorizontalSpacing(30)
+        lay.setVerticalSpacing(20)
+        lay.addWidget(lbl, 0, 0, 1, 4)
+        lay.addLayout(lay_temp1, 1, 0, 2, 1)
+        lay.addLayout(lay_temp2, 1, 1, 2, 1)
+        lay.addLayout(lay_dtemp, 1, 2, 2, 1)
+        lay.addLayout(lay_flwrt, 1, 3)
+        lay.addLayout(lay_vac, 2, 3)
+
+        self.setStyleSheet("""
+            SiriusLabel{
+                qproperty-alignment: AlignLeft;
+            }
+            QLed{
+                max-width: 1.29em;
+            }
+            .QLabel{
+                max-height:4em;
+                qproperty-alignment: AlignRight;
+            }""")
